Warning fixes : uint64 to size_t conversions.
git-svn-id: file:///home/michael/svn/yuni/trunk@2241 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/io/directory/copy.cpp
@@ -137,7 +137,7 @@ namespace Directory
 					if (toFile.open(tmp, IO::OpenMode::write | IO::OpenMode::truncate))
 					{
 						// reading the whole source file
-						size_t numRead;
+						uint64 numRead;
 						while ((numRead = fromFile.read(buffer, bufferSize)) != 0)
 						{
 							// progression
